seperate design and creative work sections into components in work folder
