+++ src/features/layout/work/creativeWork.tsx
@@ -1,0 +1,5 @@
+import styles from "./work.module.scss";
+
+export function CreativeWork() {
+  return <section className={styles.workSectionCreative}>CreativeWork</section>;
+}
